Refactor MVC controller

+++ KFM.MVCWebApp/Controllers/PondsController.cs
@@ -10,6 +10,8 @@ using KFM.Service;
 using KFM.Common;
 using Newtonsoft.Json;
 using KFM.Service.Base;
+using Azure;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace KFM.MVCWebApp.Controllers
 {
@@ -49,7 +51,24 @@ namespace KFM.MVCWebApp.Controllers
         // GET: Ponds/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            using (var httpClient = new HttpClient())
+            {
+                using (var response = await httpClient.GetAsync(Const.API_ENDPOINT + "Ponds/" + id))
+                {
+                    if (response.IsSuccessStatusCode)
+                    {
+                        var content = await response.Content.ReadAsStringAsync();
+                        var result = JsonConvert.DeserializeObject<BusinessResult>(content);//parse from api json to BusinessResult
+                        if (result != null && result.Data != null)
+                        {
+                            var data = JsonConvert.DeserializeObject<Pond>(result!.Data!.ToString()!);//parse from BusinessResult to List<Pond>
+                            return View(data);
+                        }
+                    }
+                }
+            }
+            return View(new Pond());
+            /*if (id == null)
             {
                 return NotFound();
             }
@@ -61,12 +80,13 @@ namespace KFM.MVCWebApp.Controllers
                 return NotFound();
             }
 
-            return View(pond);
+            return View(pond);*/
         }
 
         // GET: Ponds/Create
         public IActionResult Create()
         {
+            //ViewData["YourViewData"] = new SelectList(await this.GetList(), "DataId", "Lable", Entity.DataId);
             return View();
         }
 
@@ -75,21 +95,69 @@ namespace KFM.MVCWebApp.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("PondId,Name,Image,Size,Depth,Volume,DrainCount,PumpCapacity,Description,CreatedAt,UpdatedAt")] Pond pond)
+        public async Task<IActionResult> Create(/*[Bind("PondId,Name,Image,Size,Depth,Volume,DrainCount,PumpCapacity,Description,CreatedAt,UpdatedAt")]*/Pond pond)
         {
+            bool saveStatus = false;
             if (ModelState.IsValid)
+            {
+                using (var httpClient = new HttpClient())
+                {
+                    using (var response = await httpClient.PostAsJsonAsync(Const.API_ENDPOINT + "Ponds/", pond))
+                    {
+                        if (response.IsSuccessStatusCode)
+                        {
+                            var content = await response.Content.ReadAsStringAsync();
+                            var result = JsonConvert.DeserializeObject<BusinessResult>(content);//parse from api json to BusinessResult
+                            if (result != null && result.Status == Const.SUCCESS_CREATE_CODE)
+                            {
+                                saveStatus = true;
+                            }
+                        }
+                    }
+                }
+            }
+            if (saveStatus)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                //ViewData["YourViewData"] = new SelectList(await this.GetList(), "DataId", "Lable", Entity.DataId);
+                return View(pond);
+            }         
+            /*if (ModelState.IsValid)
             {
                 _context.Add(pond);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            return View(pond);
+            return View(pond);*/
         }
 
         // GET: Ponds/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            var pond = new Pond();
+            using (var httpClient = new HttpClient())
+            {
+                using (var response = await httpClient.GetAsync(Const.API_ENDPOINT + "Ponds/" + id))
+                {
+                    if (response.IsSuccessStatusCode)
+                    {
+                        var context = await response.Content.ReadAsStringAsync();
+                        var result = JsonConvert.DeserializeObject<BusinessResult>(context);
+
+                        if (result != null && result.Data != null)
+                        {
+                            // Kiểm tra xem dữ liệu có phải là một đối tượng hay không
+                            pond = JsonConvert.DeserializeObject<Pond>(result.Data.ToString());
+                        }
+                    }
+                }
+            }
+            //ViewData["YourViewData"] = new SelectList(await this.GetList(), "DataId", "Lable", Entity.DataId);
+            return View(pond);
+            /*if (id == null)
             {
                 return NotFound();
             }
@@ -99,7 +167,7 @@ namespace KFM.MVCWebApp.Controllers
             {
                 return NotFound();
             }
-            return View(pond);
+            return View(pond);*/
         }
 
         // POST: Ponds/Edit/5
@@ -109,14 +177,25 @@ namespace KFM.MVCWebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("PondId,Name,Image,Size,Depth,Volume,DrainCount,PumpCapacity,Description,CreatedAt,UpdatedAt")] Pond pond)
         {
-            if (id != pond.PondId)
-            {
-                return NotFound();
-            }
-
+            bool updateStatus = false;
             if (ModelState.IsValid)
             {
-                try
+                using (var httpClient = new HttpClient())
+                {
+                    using (var response = await httpClient.PutAsJsonAsync(Const.API_ENDPOINT + "Ponds/", pond))
+                    {
+                        if (response.IsSuccessStatusCode)
+                        {
+                            var content = await response.Content.ReadAsStringAsync();
+                            var result = JsonConvert.DeserializeObject<BusinessResult>(content);
+                            if (result != null && result.Status == Const.SUCCESS_CREATE_CODE)
+                            {
+                                updateStatus = true;
+                            }
+                        }
+                    }
+                }
+                /*try
                 {
                     _context.Update(pond);
                     await _context.SaveChangesAsync();
@@ -132,8 +211,13 @@ namespace KFM.MVCWebApp.Controllers
                         throw;
                     }
                 }
+                return RedirectToAction(nameof(Index));*/
+            }
+            if (updateStatus)
+            {
                 return RedirectToAction(nameof(Index));
             }
+            //ViewData["YourViewData"] = new SelectList(await this.GetList(), "DataId", "Lable", Entity.DataId);
             return View(pond);
         }
 
@@ -144,15 +228,24 @@ namespace KFM.MVCWebApp.Controllers
             {
                 return NotFound();
             }
-
-            var pond = await _context.Ponds
-                .FirstOrDefaultAsync(m => m.PondId == id);
-            if (pond == null)
+            using (var httpClient = new HttpClient())
             {
-                return NotFound();
-            }
+                using (var response = await httpClient.GetAsync(Const.API_ENDPOINT + "Ponds/" + id))
+                {
+                    if (response.IsSuccessStatusCode)
+                    {
+                        var context = await response.Content.ReadAsStringAsync();
+                        var result = JsonConvert.DeserializeObject<BusinessResult>(context);
 
-            return View(pond);
+                        if (result != null && result.Data != null)
+                        {
+                            var data = JsonConvert.DeserializeObject<Pond>(result.Data.ToString());
+                            return View(data);
+                        }
+                    }
+                }
+            }
+            return View(new Pond());
         }
 
         // POST: Ponds/Delete/5
@@ -160,14 +253,42 @@ namespace KFM.MVCWebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var pond = await _context.Ponds.FindAsync(id);
+            bool deleteStatus = false;
+            if (ModelState.IsValid)
+            {
+                using(var httpClient = new HttpClient())
+                {
+                    using(var response = await httpClient.DeleteAsync(Const.API_ENDPOINT + "Ponds/" + id))
+                    {
+                        if (response.IsSuccessStatusCode)
+                        {
+                            var content = await response.Content.ReadAsStringAsync();
+                            var result = JsonConvert.DeserializeObject<BusinessResult> (content);
+                            if( result != null && result.Status == Const.SUCCESS_DELETE_CODE)
+                            {
+                                deleteStatus = true;
+                            }
+                        }
+                    }
+                }
+            }
+            if (deleteStatus)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return RedirectToAction(nameof(Delete));
+            }
+ 
+            /*var pond = await _context.Ponds.FindAsync(id);
             if (pond != null)
             {
                 _context.Ponds.Remove(pond);
             }
 
             await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction(nameof(Index));*/
         }
 
         private bool PondExists(int id)
